--- packages/akar/src/APrimitive/index.ts
@@ -1,2 +1,0 @@
-export { APrimitive, type APrimitiveProps, type AsTag } from './APrimitive';
-export { default as APrimitiveButtonExample } from './APrimitiveButtonExample.vue';
