+++ packages/akar/src/a-primitive/index.ts
@@ -1,0 +1,2 @@
+export { APrimitive, type APrimitiveProps, type AsTag } from './a-primitive';
+export { default as APrimitiveButtonExample } from './a-primitive-button-example.vue';
